api_gateway/pyproject.toml
[tool.poetry]
name = "api-gateway"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.11"
environs = "^14.1.1"
aiohttp = "^3.11.13"
pyyaml = "^6.0.2"
uvicorn = "^0.34.0"
faststream = {extras = ["rabbit"], version = "^0.5.39"}
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
pydantic = {extras = ["email"], version = "^2.11.3"}
dishka = "^1.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
